Update cloud build process for kustomize.
