+++ kustomize/internal/commands/config/config.go
@@ -1,7 +1,7 @@
 // Copyright 2019 The Kubernetes Authors.
 // SPDX-License-Identifier: Apache-2.0
 
-package misc
+package config
 
 import (
 	"fmt"
