+++ releasing/cloudbuild.yaml
@@ -1,10 +1,9 @@
 steps:
-- name: "gcr.io/cloud-builders/git"
-  args: [fetch, --tags, --depth=100]
-- name: "gcr.io/kubebuilder/goreleaser_with_go_1.12.5:0.0.1"
-  args: ["bash", "releasing/cloudbuild.sh"]
+- name: goreleaser/goreleaser
+  entrypoint: /bin/sh
+  args: ["releasing/cloudbuild.sh"]
   secretEnv: ['GITHUB_TOKEN']
 secrets:
 - kmsKeyName: projects/kustomize-199618/locations/global/keyRings/github-tokens/cryptoKeys/gh-release-token
   secretEnv:
-   GITHUB_TOKEN: CiQAyrREbPgXJOeT7M3t+WlxkhXwlMPudixBeiyWTjmLOMLqdK4SUQA0W+xUmDJKAhyfHCcwqSEzUn9OwKC7XAYcmwe0CCKTCbPbDgmioDK24q3LVapndXNvnnHvCjhOJNEr1o+P1DCF+LlzYV2YL8lP09rrKrslPg==
+    GITHUB_TOKEN: CiQAyrREbPgXJOeT7M3t+WlxkhXwlMPudixBeiyWTjmLOMLqdK4SUQA0W+xUmDJKAhyfHCcwqSEzUn9OwKC7XAYcmwe0CCKTCbPbDgmioDK24q3LVapndXNvnnHvCjhOJNEr1o+P1DCF+LlzYV2YL8lP09rrKrslPg==
